resolve cyclic dependency by extracting test activity calender into new file
