+++ frontend/__tests__/elements/test-activity-calendar.spec.ts
@@ -1,17 +1,16 @@
 import {
   TestActivityCalendar,
-  ActivityDay,
   ModifiableTestActivityCalendar,
-} from "../../src/ts/elements/test-activity";
+} from "../../src/ts/elements/test-activity-calendar";
 import * as Dates from "date-fns";
 import { MatcherResult } from "../vitest";
+import { UTCDateMini } from "@date-fns/utc/date/mini";
 
-describe("test-activity.ts", () => {
+describe("test-activity-calendar.ts", () => {
   describe("TestActivityCalendar", () => {
     describe("getMonths", () => {
       it("for lastDay in april", () => {
         const calendar = new TestActivityCalendar([], getDate("2024-04-10"));
-        console.log(Dates.format(getDate("2024-04-10"), "yyyy-MM-dd hh:mm:ss"));
 
         expect(calendar.getMonths()).toEqual([
           "May`23",
@@ -153,7 +152,6 @@ describe("test-activity.ts", () => {
     it("for full year", () => {
       //GIVEN
       const data = getData("2022-11-30", "2023-12-31");
-      console.log(new Date("2023-12-31T23:59:59Z").valueOf());
       const calendar = new TestActivityCalendar(
         data,
         new Date("2023-12-31T23:59:59Z")
@@ -347,8 +345,7 @@ describe("test-activity.ts", () => {
 });
 
 function getDate(date: string): Date {
-  const parts = date.split("-").map((it) => parseInt(it));
-  return new Date(Date.UTC(parts[0], parts[1] - 1, parts[2], 0, 0, 0, 0));
+  return new UTCDateMini(Dates.parseISO(date + "T00:00:00Z"));
 }
 
 function getData(from: string, to: string): number[] {
