+++ frontend/src/ts/elements/test-activity.ts
@@ -1,178 +1,9 @@
 import SlimSelect from "slim-select";
 import type { DataObjectPartial } from "slim-select/dist/store";
-import * as Dates from "date-fns";
 import { getTestActivityCalendar } from "../db";
-import { UTCDateMini } from "@date-fns/utc/date/mini";
 
 let yearSelector: SlimSelect | undefined;
 
-export type ActivityDay = {
-  level: string;
-  label?: string;
-};
-
-export class TestActivityCalendar implements MonkeyTypes.TestActivityCalendar {
-  protected data: (number | null | undefined)[];
-  protected startDay: Date;
-  protected endDay: Date;
-
-  constructor(data: (number | null)[], lastDay: Date) {
-    const local = new UTCDateMini(lastDay);
-    console.log({
-      offset: new Date().getTimezoneOffset() / 60,
-      utc: lastDay.valueOf(),
-      local: local.valueOf(),
-      utcFormat: Dates.format(lastDay, "yyyy-MMM-dd HH:mm:ss"),
-      localFormat: Dates.format(local, "yyyy-MMM-dd HH:mm:ss"),
-    });
-    const interval = this.getInterval(local);
-    this.startDay = interval.start as Date;
-    this.endDay = interval.end as Date;
-    this.data = this.buildData(data, local);
-  }
-
-  protected getInterval(lastDay: Date): Dates.Interval {
-    const end = Dates.endOfMonth(lastDay);
-    const start = Dates.addDays(Dates.subYears(end, 1), 1);
-    /*
-    console.log({
-      offset: new Date().getTimezoneOffset() / 60,
-      lastDay: Dates.format(lastDay, "EEEE dd MMM yyyy"),
-      lastDayValue: lastDay.valueOf(),
-      end: Dates.format(end, "EEEE dd MMM yyyy"),
-      start: Dates.format(start, "EEEE dd MMM yyyy"),
-    });
-    */
-    return { start, end };
-  }
-
-  protected buildData(
-    data: (number | null | undefined)[],
-    lastDay: Date
-  ): (number | null | undefined)[] {
-    //fill calendar with enough values
-    const values = new Array(Math.max(0, 386 - data.length)).fill(undefined);
-    values.push(...data);
-
-    //discard values outside the calendar range
-    const days = Dates.differenceInDays(this.endDay, this.startDay) + 1;
-    const offset =
-      values.length - days + Dates.differenceInDays(this.endDay, lastDay);
-    return values.slice(offset);
-  }
-
-  getMonths(): string[] {
-    return Dates.eachMonthOfInterval({
-      start: this.startDay,
-      end: this.endDay,
-    }).map((month) => Dates.format(month, "MMM`yy"));
-  }
-
-  getDays(): ActivityDay[] {
-    const result: ActivityDay[] = [];
-
-    const buckets = this.getBuckets();
-    const getValue = (v: number | null | undefined): string => {
-      if (v === undefined) return "filler";
-      if (v === null || v === 0) return "0";
-      for (let b = 0; b < 4; b++)
-        if (v <= (buckets[b] ?? 0)) return (1 + b).toString();
-
-      return "4";
-    };
-
-    //skip weekdays in the previous month
-    for (let i = 0; i < this.startDay.getDay(); i++) {
-      result.push({
-        level: "filler",
-      });
-    }
-
-    const days = Dates.differenceInDays(this.endDay, this.startDay) + 1;
-
-    let currentDate = this.startDay;
-    for (let i = 0; i < days; i++) {
-      const count = this.data[i];
-      /*
-      console.log({
-        i,
-        count,
-        date: Dates.format(currentDate, "EEEE dd MMM yyyy"),
-      });
-      */
-      result.push({
-        level: getValue(count),
-        label:
-          this.data[i] !== undefined && this.data[i] !== null
-            ? `${count} ${count == 1 ? "test" : "tests"} on ${Dates.format(
-                currentDate,
-                "EEEE dd MMM yyyy"
-              )}`
-            : undefined,
-      });
-      currentDate = Dates.addDays(currentDate, 1);
-    }
-
-    //add weekdays missing
-    for (let i = this.endDay.getDay(); i < 6; i++) {
-      result.push({
-        level: "filler",
-      });
-    }
-
-    return result;
-  }
-
-  private getBuckets(): number[] {
-    const filtered = this.data.filter(
-      (it) => it !== null && it !== undefined
-    ) as number[];
-    const sorted = filtered.sort((a, b) => a - b);
-
-    const trimmed = sorted.slice(
-      Math.round(sorted.length * 0.1),
-      sorted.length - Math.round(sorted.length * 0.1)
-    );
-    const sum = trimmed.reduce((a, c) => a + c, 0);
-    const mid = sum / trimmed.length;
-    return [Math.floor(mid / 2), Math.round(mid), Math.round(mid * 1.5)];
-  }
-}
-
-export class ModifiableTestActivityCalendar
-  extends TestActivityCalendar
-  implements MonkeyTypes.ModifiableTestActivityCalendar
-{
-  private lastDay: Date;
-
-  constructor(data: (number | null)[], lastDay: Date) {
-    super(data, lastDay);
-    this.lastDay = lastDay;
-  }
-
-  increment(date: Date): void {
-    //TODO timezones?
-    if (Dates.isSameDay(date, this.lastDay)) {
-      const last = this.data.length - 1;
-      this.data[last] = (this.data[last] || 0) + 1;
-    } else if (Dates.isBefore(date, this.lastDay)) {
-      throw new Error("cannot alter data in the past.");
-    } else {
-      const missedDays = Dates.differenceInDays(date, this.lastDay) - 1;
-      for (let i = 0; i < missedDays; i++) {
-        this.data.push(undefined);
-      }
-      this.data.push(1);
-      //update timeframe
-      const interval = this.getInterval(date);
-      this.startDay = interval.start as Date;
-      this.endDay = interval.end as Date;
-      this.lastDay = date;
-    }
-
-    this.data = this.buildData(this.data, this.lastDay);
-  }
-}
 export function init(calendar?: MonkeyTypes.TestActivityCalendar): void {
   if (calendar === undefined) {
     $("#testActivity").addClass("hidden");
